+++ lib/presentation/pages/home/home_page.dart
@@ -8,7 +8,7 @@ import 'package:techtalk/core/theme/extension/app_color.dart';
 import 'package:techtalk/presentation/pages/home/widgets/cheer_up_message_card.dart';
 import 'package:techtalk/presentation/pages/home/widgets/home_state.dart';
 import 'package:techtalk/presentation/pages/home/widgets/practical_interview_card.dart';
-import 'package:techtalk/presentation/pages/home/widgets/topic_interview_card.dart';
+import 'package:techtalk/presentation/pages/home/widgets/single_topic_interview_card.dart';
 import 'package:techtalk/presentation/widgets/base/base_page.dart';
 
 class HomePage extends BasePage with HomeState {
@@ -22,13 +22,13 @@ class HomePage extends BasePage with HomeState {
       data: (_) {
         return ListView(
           physics: const ScrollPhysics(),
-          padding: const EdgeInsets.symmetric(vertical: 8),
+          padding: const EdgeInsets.symmetric(vertical: 8, horizontal: 16),
           children: const [
             CheerUpMessageCard(),
             Gap(16),
             PracticalInterviewCard(),
             Gap(16),
-            TopicInterviewCard(),
+            SingleTopicInterviewCard(),
           ],
         );
       },
