+++ lib/presentation/pages/home/widgets/practical_interview_card.dart
@@ -1,56 +1,60 @@
 import 'package:flutter/material.dart';
-import 'package:font_awesome_flutter/font_awesome_flutter.dart';
-import 'package:gap/gap.dart';
+import 'package:flutter_svg/flutter_svg.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
+import 'package:techtalk/core/constants/assets.dart';
+import 'package:techtalk/core/constants/interview_type.enum.dart';
 import 'package:techtalk/core/theme/extension/app_color.dart';
 import 'package:techtalk/core/theme/extension/app_text_style.dart';
 import 'package:techtalk/presentation/pages/home/home_event.dart';
+import 'package:techtalk/presentation/pages/home/widgets/home_state.dart';
 
-class PracticalInterviewCard extends ConsumerWidget with HomeEvent {
+class PracticalInterviewCard extends ConsumerWidget with HomeState, HomeEvent {
   const PracticalInterviewCard({super.key});
 
   @override
   Widget build(BuildContext context, WidgetRef ref) {
-    return Padding(
-      padding: const EdgeInsets.symmetric(horizontal: 16),
-      child: Material(
-        clipBehavior: Clip.antiAlias,
-        color: AppColor.of.brand1,
-        borderRadius: BorderRadius.circular(16),
-        child: InkWell(
-          onTap: () => onPracticalCardTapped(ref),
-          child: Container(
-            padding: const EdgeInsets.all(24),
-            child: Column(
-              crossAxisAlignment: CrossAxisAlignment.start,
-              children: [
-                Row(
-                  children: [
-                    Expanded(
-                      child: Text(
-                        '실전형 면접',
-                        style: AppTextStyle.headline2.copyWith(
-                          color: AppColor.of.brand3,
-                        ),
+    return Material(
+      clipBehavior: Clip.antiAlias,
+      color: AppColor.of.brand1,
+      borderRadius: BorderRadius.circular(16),
+      child: InkWell(
+        onTap: () => onPracticalCardTapped(ref),
+        child: Container(
+          padding: const EdgeInsets.fromLTRB(24, 12, 0, 12),
+          child: Column(
+            crossAxisAlignment: CrossAxisAlignment.start,
+            children: [
+              Row(
+                mainAxisAlignment: MainAxisAlignment.spaceBetween,
+                children: [
+                  Expanded(
+                    child: Text(
+                      '실전형 면접',
+                      style: AppTextStyle.headline2.copyWith(
+                        color: AppColor.of.brand3,
                       ),
                     ),
-                    const Gap(48),
-                    FaIcon(
-                      FontAwesomeIcons.circlePlus,
-                      color: AppColor.of.brand2,
-                      size: 24,
+                  ),
+                  GestureDetector(
+                    onTap: () {
+                      routeToTopicSelectPage(context,
+                          type: InterviewType.practical);
+                    },
+                    child: SvgPicture.asset(Assets.iconsRoundBlueCircle),
+                  ),
+                ],
+              ),
+              if (targetedTopics(ref).isEmpty)
+                Padding(
+                  padding: const EdgeInsets.only(bottom: 12),
+                  child: Text(
+                    '여러 주제를 선택해 실전 연습을 해보세요!',
+                    style: AppTextStyle.body1.copyWith(
+                      color: AppColor.of.gray3,
                     ),
-                  ],
-                ),
-                const Gap(12),
-                Text(
-                  '여러 주제를 선택해 실전 연습을 해보세요!',
-                  style: AppTextStyle.body1.copyWith(
-                    color: AppColor.of.gray3,
                   ),
                 ),
-              ],
-            ),
+            ],
           ),
         ),
       ),
